script 5

+++ recursion/main.h
@@ -10,4 +10,4 @@ int _strlen_recursion(char *s);
 int factorial(int n);
 int _pow_recursion(int x, int y);
 int _sqrt_recursion(int n);
-int square(int x, int y);
+int square(unsigned long x, unsigned long y);
